requirements.txt
python-dotenv==0.15.0
requests==2.25.1

